+++ luna-ui/react-native/src/theme/Theme.ts
@@ -1,0 +1,9 @@
+import { Colors } from "./colors";
+import { Spacing } from "./spacing";
+import { Typography } from "./typography";
+
+export interface LunaTheme {
+  colors: Colors;
+  spacing: Spacing;
+  typography: Typography;
+}
