+++ luna-ui/react-native/src/theme/colors.ts
@@ -1,32 +1,25 @@
-export const LunaLightThemeColors = {
-  primary: "#005EB8",
-  secondary: "#41B6E6",
-  lightText: "#FFFFFF",
-  darkText: "#000000",
+export type Colors = {
+  primary: string;
+  secondary: string;
 
-  // Main Colours
-  darkAccent: "#003087",
-  brightAccent: "#0072CE",
-  lightAccent: "#41B6E6",
-  aquaAccent: "#00A9CE",
+  text: string;
+  background: string;
+  surface: string;
 
-  // Supporting Colours
-  darkSupport: "#006747",
-  support: "#009639",
-  lightSupport: "#78BE20",
-  aquaSupport: "#00A499",
+  // Main Colors
+  darkAccent: string;
+  brightAccent: string;
+  lightAccent: string;
+  aquaAccent: string;
 
-  // Neutral colours
-  black: "#231f20",
-  darkGrey: "#425563",
-  midGrey: "#768692",
-  paleGrey: "#E8EDEE",
-  white: "white",
+  // Supporting Colors
+  darkSupport: string;
+  support: string;
+  lightSupport: string;
+  aquaSupport: string;
 
-  // Alerts Colours
-  error: "#DA291C",
-  success: "#78BE20",
-
-  // static const warning :"#FFB81C";
-  warning: "#FAE100",
-} as const;
+  // Alerts Colors
+  error: string;
+  success: string;
+  warning: string;
+};
