+++ luna-ui/react-native/src/theme/index.ts
@@ -2,3 +2,5 @@ export * from "./colors";
 export * from "./typography";
 export * from "./spacing";
 export * from "./useTheme";
+export * from "./Theme";
+export * from "./ThemeProvider";
