+++ luna-ui/react-native/src/theme/spacing.ts
@@ -1,18 +1,10 @@
 export type LunaSpacing = "xs" | "sm" | "md" | "lg" | "xl" | number;
-
-export const LunaSpacingMap = {
-  xs: 4,
-  sm: 8,
-  md: 16,
-  lg: 24,
-  xl: 32,
-} as const;
-
-// export const Spacing = {
-//   extraSmall: 8.0,
-//   small: 12.0,
-//   medium: 16.0,
-//   large: 24.0,
-//   extraLarge: 32.0,
-//   xxlLarge: 48.0,
-// } as const;
+export type Spacing = {
+  none: number;
+  xs: number;
+  sm: number;
+  md: number;
+  lg: number;
+  xl: number;
+  xxl: number;
+};
